refactor backend to model package
