+++ src/main/java/org/cryptomator/hubcli/Backend.java
@@ -1,9 +1,8 @@
 package org.cryptomator.hubcli;
 
-import com.fasterxml.jackson.annotation.JsonProperty;
-import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.cryptomator.hubcli.model.UserDto;
+import org.cryptomator.hubcli.model.VaultDto;
 
 import java.io.IOException;
 import java.net.URI;
@@ -12,9 +11,7 @@ import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
 import java.nio.charset.StandardCharsets;
 import java.time.Duration;
-import java.time.Instant;
 import java.util.Arrays;
-import java.util.Objects;
 import java.util.UUID;
 
 public class Backend implements AutoCloseable {
@@ -25,9 +22,9 @@ public class Backend implements AutoCloseable {
     private final VaultService vaultService;
     private final UserService userService;
 
-    public Backend(AccessToken token, Common common) {
-        this.accessToken = token.value;
-        this.apiBase = common.getApiBase();
+    public Backend(String accessToken, URI apiBase) {
+        this.accessToken = accessToken;
+        this.apiBase = apiBase;
 
         this.vaultService = new VaultService();
         this.userService = new UserService();
@@ -44,7 +41,8 @@ public class Backend implements AutoCloseable {
     class VaultService {
 
         public HttpResponse<String> createOrUpdateVault(UUID vaultId, String name, String description, boolean archived) throws IOException, InterruptedException, UnexpectedStatusCodeException {
-            var vault = new VaultDto(vaultId, name, Objects.requireNonNullElse(description, ""), archived, Instant.now().toString(), "asd", 3, "asd", "asd", "asd"); //TODO: backend allows nullable description, but frontend is broken
+            //creationTime is ignored
+            var vault = new VaultDto(vaultId, name, description, archived, "1970-01-01T00:00:00Z", null, 0, null, null, null);
             var req = createRequest("vaults/" + vaultId).PUT(HttpRequest.BodyPublishers.ofString(vault.toJson())).build();
             return sendRequest(HttpClient.newHttpClient(), req, HttpResponse.BodyHandlers.ofString(StandardCharsets.UTF_8), 200, 201);
         }
@@ -52,26 +50,6 @@ public class Backend implements AutoCloseable {
         public HttpResponse<String> grantAccess(UUID vaultId, String userId, String jwe) throws IOException, InterruptedException, UnexpectedStatusCodeException {
             var req = createRequest("vaults/" + vaultId + "/access-tokens/" + userId).PUT(HttpRequest.BodyPublishers.ofString(jwe)).header("Content-Type", "text/plain").build();
             return sendRequest(HttpClient.newHttpClient(), req, HttpResponse.BodyHandlers.ofString(StandardCharsets.UTF_8), 201);
-        }
-
-        public record VaultDto(@JsonProperty("id") UUID id,
-                               @JsonProperty("name") String name,
-                               @JsonProperty("description") String description,
-                               @JsonProperty("archived") boolean archived,
-                               @JsonProperty("creationTime") String creationTime,
-                               //TODO: To use instant we need additional jackson library
-                               // Legacy properties ("Vault Admin Password"):
-                               @JsonProperty("masterkey") String masterkey,
-                               @JsonProperty("iterations") Integer iterations,
-                               @JsonProperty("salt") String salt,
-                               @JsonProperty("authPublicKey") String authPublicKey,
-                               @JsonProperty("authPrivateKey") String authPrivateKey
-        ) {
-
-            public String toJson() throws JsonProcessingException {
-                return new ObjectMapper().writeValueAsString(this);
-            }
-
         }
 
     }
